daprApps_v1/proxy/requirements.txt
dapr-ext-grpc-dev >= 1.8.0rc1.dev
dapr-dev >= 1.8.0rc1.dev
prometheus-client
numpy